Aqui eu comecei o frontend, e também comecei algumas adaptações no backend para o melhor funcionamento da aplicação.

+++ backend/api/src/main/java/acad/reddragon/api/detalhe_treino/DetalheTreinoRepository.java
@@ -1,5 +1,7 @@
 package acad.reddragon.api.detalhe_treino;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalheTreinoRepository  extends JpaRepository<DetalheTreino, Long> {
+import java.math.BigInteger;
+
+public interface DetalheTreinoRepository  extends JpaRepository<DetalheTreino, BigInteger> {
 }
